+++ lib/core/files/generate_page_file.dart
@@ -1,16 +1,19 @@
 import 'dart:io';
 
-import 'package:js_cli/core/files/configs_file.dart';
+import 'package:js_cli/core/utils/directory_utils.dart';
+import 'package:js_cli/core/utils/reserved_words.dart';
+
+import 'configs_file.dart';
 
 class GeneratePageFile {
-  static String templeteNone = r'''
+  static String template = r'''
 import 'package:flutter/material.dart';
 import '{{controllerNameFile.snakeCase}}.dart';
 
-class {{name.pascalCase}} extends StatefulWidget {
+class {{pageNameClass.pascalCase}} extends StatefulWidget {
 
   @override
-  _{{name.pascalCase}}State createState() => _{{name.pascalCase}}State();
+  _{{pageNameClass.pascalCase}}State createState() => _{{pageNameClass.pascalCase}}State();
 }
 
 class _{{name.pascalCase}}State extends State<{{pageNameClass.pascalCase}}> {
@@ -27,54 +30,34 @@ class _{{name.pascalCase}}State extends State<{{pageNameClass.pascalCase}}> {
 }
           
           ''';
-  static String templeteFlutterModular = r'''
-import 'package:flutter/material.dart';
-import 'package:flutter_modular/flutter_modular.dart';
-
-import '{{controllerNameFile.snakeCase}}.dart';
-
-class {{name.pascalCase}} extends StatefulWidget {
-
-  @override
-  _{{name.pascalCase}}State createState() => _{{name.pascalCase}}State();
-}
-
-class _{{name.pascalCase}}State extends ModularState<{{pageNameClass.pascalCase}}, {{controllerNameClass.pascalCase}}> {
-  final GlobalKey<ScaffoldState> scaffoldKey = new GlobalKey<ScaffoldState>();
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      key: scaffoldKey,
-      appBar: AppBar(),
-      body: Container(),
+  static String readImp(String name) {
+    return _read(
+      name,
+      'page.template',
+      template,
+      ReservedWords.removeWordsInFile(
+        fileString: ConfigsFile.getPagePath(),
+      ),
     );
   }
-}
-          
-          ''';
 
-  static String read() {
-    final integration = ConfigsFile.getIntegration();
-    var path;
-    var templete;
-    if (integration == 'flutter_modular') {
-      path = 'generate_flutter_modular_page.template';
-      templete = templeteFlutterModular;
-    } else {
-      path = 'generate_none_page.template';
-      templete = templeteNone;
-    }
+  static String _read(
+    String name,
+    String file,
+    String template,
+    String path,
+  ) {
+    var root = '.js_cli';
 
-    var existFile = File('.js_cli/template/$path').existsSync();
+    path = 'template/$path';
+
+    var existFile = File('$root/$path/$file').existsSync();
 
     if (!existFile) {
-      Directory('.js_cli').createSync();
-      Directory('.js_cli/template').createSync();
-  
-      File('.js_cli/template/$path').writeAsStringSync(templete);
+      DirectoryUtils.create(root, path);
+      File('$root/$path/$file').writeAsStringSync(template);
     }
 
-    return File('.js_cli/template/$path').readAsStringSync();
+    return File('$root/$path/$file').readAsStringSync();
   }
 }