+++ lib/core/interfaces/igenerate_datasources.dart
@@ -2,12 +2,13 @@ abstract class IGenerateDatasources {
   Future<bool> call({
     required String name,
     required String path,
+    required String current,
   });
 
-  String getNameFile(String name);
-  String getNameFileInterface(String name);
-  String getNameClass(String name);
-  String getNameClassInterface(String name);
+  String getNameFile(String name, String current);
+  String getNameFileInterface(String name, String current);
+  String getNameClass(String name, String current);
+  String getNameClassInterface(String name, String current);
   String getPath();
   String getPathInterface();
 }