+++ lib/core/template_default/start/folder_structure.dart
@@ -1,0 +1,26 @@
+class FolderStruture {
+  final template = [
+    {
+      'data': [
+        'datasources',
+        'repositories',
+      ],
+    },
+    {
+      'domain': [
+        {
+          'models': [
+            'dtos',
+            'entities',
+          ],
+        },
+        'repositories',
+        'usecases',
+      ],
+    },
+    {
+      'external': [],
+    },
+    'presentation',
+  ];
+}