+++ lib/core/utils/directory_utils.dart
@@ -9,6 +9,7 @@ class DirectoryUtils {
     var complete = '$root';
     for (var i = 0; i < path.split('/').length; i++) {
       final e = path.split('/')[i];
+      if (e.isEmpty) continue;
       complete += '/$e';
 
       final exist = Directory(complete).existsSync();