+++ lib/modules/generate_layers/usecases/domain/generate_controller.dart
@@ -3,8 +3,8 @@ import 'dart:io';
 import 'package:js_cli/core/files/configs_file.dart';
 import 'package:js_cli/core/files/generate_controller_file.dart';
 import 'package:js_cli/core/interfaces/igenerate_controller.dart';
-import 'package:js_cli/core/templates/core/generic_template.dart';
 import 'package:js_cli/core/utils/directory_utils.dart';
+import 'package:js_cli/core/utils/replace_utils.dart';
 import 'package:js_cli/core/utils/reserved_words.dart';
 
 import '../../../../core/errors/file_exists_error.dart';
@@ -14,18 +14,22 @@ class GenerateController implements IGenerateController {
   Future<bool> call({
     required String name,
     required String path,
+    required String current,
   }) async {
     DirectoryUtils.create(
       path,
       ReservedWords.replaceWordsInFile(
-        fileString: getPath(),
+        fileString: getPath(name, current),
         name: name,
+        current: current,
       ),
     );
 
     var completePath = ReservedWords.replaceWordsInFile(
-      fileString: '$path/${getPath()}/${getNameFile(name)}.dart',
+      fileString:
+          '$path/${getPath(name, current)}/${getNameFile(name, current)}.dart',
       name: name,
+      current: current,
     );
 
     if (File(completePath).existsSync()) {
@@ -33,39 +37,49 @@ class GenerateController implements IGenerateController {
     }
 
     String? content = ReservedWords.replaceWordsInFile(
-      fileString: GenerateControllerFile.read(),
+      fileString: GenerateControllerFile.readImp(name),
       name: name,
       path: path,
+      current: current,
     );
 
     File(completePath).writeAsStringSync(content);
 
-    updateIntegrationModule(
-      path,
-      getNameClass(name),
-      getPath(),
+    applyReplaceIfNecessary(
+      current: current,
+      name: name,
+      path: path,
+      subPath: ConfigsFile.getControllerPath(),
+      prefixNameReplaceFile: 'controller',
     );
+
     return true;
   }
 
   @override
-  String getNameClass(String name) {
+  String getNameClass(String name, String current) {
     return ReservedWords.replaceWordsInFile(
       fileString: ConfigsFile.getControllerNameClass(),
       name: name,
+      current: current,
     );
   }
 
   @override
-  String getNameFile(String name) {
+  String getNameFile(String name, String current) {
     return ReservedWords.replaceWordsInFile(
       fileString: ConfigsFile.getControllerNameFile(),
       name: name,
+      current: current,
     );
   }
 
   @override
-  String getPath() {
-    return ConfigsFile.getPagePath();
+  String getPath(String name, String current) {
+    return ReservedWords.replaceWordsInFile(
+      fileString: ConfigsFile.getControllerPath(),
+      name: name,
+      current: current,
+    );
   }
 }