+++ lib/modules/generate_layers/usecases/domain/generate_page.dart
@@ -3,8 +3,8 @@ import 'dart:io';
 import 'package:js_cli/core/files/configs_file.dart';
 import 'package:js_cli/core/files/generate_page_file.dart';
 import 'package:js_cli/core/interfaces/igenerate_page.dart';
-import 'package:js_cli/core/templates/core/generic_template.dart';
 import 'package:js_cli/core/utils/directory_utils.dart';
+import 'package:js_cli/core/utils/replace_utils.dart';
 import 'package:js_cli/core/utils/reserved_words.dart';
 
 import '../../../../core/errors/file_exists_error.dart';
@@ -14,18 +14,22 @@ class GeneratePages implements IGeneratePages {
   Future<bool> call({
     required String name,
     required String path,
+    required String current,
   }) async {
     DirectoryUtils.create(
       path,
       ReservedWords.replaceWordsInFile(
-        fileString: getPath(),
+        fileString: getPath(name, current),
         name: name,
+        current: current,
       ),
     );
 
     var completePath = ReservedWords.replaceWordsInFile(
-      fileString: '$path/${getPath()}/${getNameFile(name)}.dart',
+      fileString:
+          '$path/${getPath(name, current)}/${getNameFile(name, current)}.dart',
       name: name,
+      current: current,
     );
 
     if (File(completePath).existsSync()) {
@@ -33,34 +37,49 @@ class GeneratePages implements IGeneratePages {
     }
 
     String? content = ReservedWords.replaceWordsInFile(
-      fileString: GeneratePageFile.read(),
+      fileString: GeneratePageFile.readImp(name),
       name: name,
       path: path,
+      current: current,
     );
 
     File(completePath).writeAsStringSync(content);
+
+    applyReplaceIfNecessary(
+      current: current,
+      name: name,
+      path: path,
+      subPath: ConfigsFile.getPagePath(),
+      prefixNameReplaceFile: 'page',
+    );
 
     return true;
   }
 
   @override
-  String getNameClass(String name) {
+  String getNameClass(String name, String current) {
     return ReservedWords.replaceWordsInFile(
       fileString: ConfigsFile.getPageNameClass(),
       name: name,
+      current: current,
     );
   }
 
   @override
-  String getNameFile(String name) {
+  String getNameFile(String name, String current) {
     return ReservedWords.replaceWordsInFile(
       fileString: ConfigsFile.getPageNameFile(),
       name: name,
+      current: current,
     );
   }
 
   @override
-  String getPath() {
-    return ConfigsFile.getPagePath();
+  String getPath(String name, String current) {
+    return ReservedWords.replaceWordsInFile(
+      fileString: ConfigsFile.getPagePath(),
+      name: name,
+      current: current,
+    );
   }
 }