+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2+1 - generate template in .js_cli
+
+-   Now the code generation will generate the file and .js_cli where it can be edited as needed.
+
 ## 0.0.1 - First Publication
 
 - Publish package